Add settings modal with sorting, app version, license and build branch
Move the dashboard customize/sort UI into a Settings modal (gear icon)
that also surfaces the app version/build, the MIT license, and the git
branch and commit the app was built from. A build phase embeds the
branch and commit into Info.plist at build time.
